+++ scripts/get_latest_baselines.py
@@ -14,8 +14,9 @@ import os
 import botocore.exceptions
 import yaml
 
-from .data_utils import get_config_params
-from .global_variables import s3
+# from .data_utils import get_config_params
+# from .global_variables import s3
+from scripts_config import get_config_params, s3
 
 
 def get_latest_baselines():
