+++ scripts/get_latest_insar_pairs.py
@@ -11,8 +11,9 @@ import os
 import botocore.exceptions
 import yaml
 
-from .data_utils import get_config_params
-from .global_variables import s3
+# from .data_utils import get_config_params
+# from .global_variables import s3
+from scripts_config import get_config_params, s3
 
 
 def get_latest_insar_pairs():
