+++ scripts/global_variables.py
@@ -1,15 +1,15 @@
-#!/usr/bin/python3
-"""
-Volcano InSAR Interpretation Workbench
+# #!/usr/bin/python3
+# """
+# Volcano InSAR Interpretation Workbench
 
-SPDX-License-Identifier: MIT
+# SPDX-License-Identifier: MIT
 
-Copyright (C) 2021-2023 Government of Canada
+# Copyright (C) 2021-2023 Government of Canada
 
-Authors:
-  - Chloe Lam <chloe.lam@nrcan-rncan.gc.ca>
-"""
+# Authors:
+#   - Chloe Lam <chloe.lam@nrcan-rncan.gc.ca>
+# """
 
-import boto3
-# verify = False for when working in office to bypass SSL Certificate Error
-s3 = boto3.client('s3', verify=False)
+# import boto3
+# # verify = False for when working in office to bypass SSL Certificate Error
+# s3 = boto3.client('s3', verify=False)
