+++ scripts/load_updated_coherence_matrix.py
@@ -12,9 +12,9 @@ Authors:
 import argparse
 import os
 
-from .data_utils import get_config_params
-from .global_variables import s3
-
+# from .data_utils import get_config_params
+# from .global_variables import s3
+from scripts_config import get_config_params, s3
 
 def main():
     """
